Massive update of build_ghrelease
Do cross-compilation "native" (no Docker image necessary)

We set up a dep area in build-deps for OpenSSL. We build
OpenSSL in that dep area and then we point openssl-sys
(via Cargo cross-compilation) to that dep area.

This adds dependencies to cross-compilation for x86_64-linux-musl,
which on macOS is quickly resolved by using:

brew install FiloSottile/musl-cross/musl-cross

We also add a dependency to rpm2cpio, because we pull the
kernel-headers package for Amazon Linux -- we need them
to compile OpenSSL for Linux, which we need to compile
openssl-sys for Linux (and specifically musl to be honest).

Lots of messaging and things have been added around what
this script is doing, and it almost is supposed to look like
a `./configure` script in many ways.

I would have continued using Docker, but the process kept
taking absolutely forever (probably because of x86_64 emulation)
and not working quite right.

TODO: support Linux as well as macOS